Renamed index.js to descriptive filenames
